drm/i915: Share local memory object with virtio-GPU
When we see a virtio-GPU upon DMA-BUF map, we can hand over device memory
addresses (offsets) so that virtio-GPU back-end in the host side is able
to construct a GEM object back from these addresses with the deviceptr
ioctl and uses it as scan-out buffer to achieve virtualized display in a
zero-copy way.

We try to share a local memory object with virtio-GPU only when
peer2peer flag is set in the DMA-BUF attachment object.  In this way we
make local memory sharing explicit to allow migration to system memory
when virtio-GPU doesn't support peer2peer (e.g., the virtio-GPU device
is backed by an iGPU in the host).

Tracked-On: OAM-117994
Signed-off-by: Weifeng Liu <weifeng.liu@intel.com>